+++ src/libtrx/game/ui/dialogs/setting_helpers/handlers.h
@@ -15,6 +15,7 @@ bool UI_Settings_EnableExitFadeEffects_IsAvailable(
 bool UI_Settings_PauseMusicInInventory_IsAvailable(
     const UI_SETTINGS_OPTION *option);
 
+bool UI_Settings_FogColor_IsAvailable(const UI_SETTINGS_OPTION *option);
 bool UI_Settings_EnableBreeze_IsAvailable(const UI_SETTINGS_OPTION *option);
 bool UI_Settings_UIPickupsScale_IsAvailable(const UI_SETTINGS_OPTION *option);
 bool UI_Settings_Healthbar_IsAvailable(const UI_SETTINGS_OPTION *option);